+++ bindings/pyroot_experimental/pythonizations/CMakeLists.txt
@@ -10,6 +10,8 @@
 
 if(dataframe)
     list(APPEND PYROOT_EXTRA_PYSOURCE
+        ROOT/pythonization/utils/__init__.py
+        ROOT/pythonization/utils/_rdf_utils.py
         ROOT/pythonization/_rdataframe.py
         ROOT/pythonization/_rtensor.py)
     list(APPEND PYROOT_EXTRA_SOURCE
