+++ bindings/pyroot_experimental/pythonizations/python/ROOT/pythonization/utils/__init__.py
@@ -1,0 +1,9 @@
+# Author: Enric Tejedor, Stefan Wunsch CERN  03/2020
+
+################################################################################
+# Copyright (C) 1995-2018, Rene Brun and Fons Rademakers.                      #
+# All rights reserved.                                                         #
+#                                                                              #
+# For the licensing terms see $ROOTSYS/LICENSE.                                #
+# For the list of contributors see $ROOTSYS/README/CREDITS.                    #
+################################################################################
